chore(eval): add ruff + mypy dev tooling; generator polish
- pyproject.toml on both packages: ruff (>=0.6,<1) and mypy (>=1.10,<3)
  as dev extras. pytest now pinned <9 (don't let CI float onto an
  unreviewed major). Per-package py.typed in package-data so mypy
  follows annotations across the editable-install boundary.

- Ruff config: line-length 100, target py311, narrow rule set
  (E/W/F/I/B/UP/SIM). isort combine-as-imports keeps aliased imports
  on one line.

- Mypy config: strict mode, with module overrides for fitz/reportlab/PIL
  (no shipped stubs) so the ignore is centralized rather than sprinkled.

- Generator polish: small typing/import cleanups across docs/ and
  scan_artifacts.py; no behavior change. Tests + manual packet
  regeneration unchanged.

+++ evals/generators/packetready_eval/docs/_layout.py
@@ -7,8 +7,8 @@ not chrome — extractor robustness to chrome variation is a P4 problem.
 
 from __future__ import annotations
 
+from collections.abc import Iterable
 from dataclasses import dataclass
-from typing import Iterable
 
 from reportlab.lib import colors
 from reportlab.lib.pagesizes import LETTER
@@ -65,7 +65,9 @@ def draw_header(c: Canvas, h: HeaderSpec) -> float:
     c.drawString(badge_x + 0.2 * inch, badge_y + 0.075 * inch, h.badge)
 
     c.setFillColor(colors.black)
-    return band_top - band_height - 0.3 * inch
+    # Cast because `inch` from unstubbed reportlab makes the arithmetic
+    # produce Any; mypy-strict wants the return type narrowed.
+    return float(band_top - band_height - 0.3 * inch)
 
 
 def draw_field_grid(
